refine ethash to make it ready for connection
