+++ consensus/ethash/blocknonce.go
@@ -7,6 +7,7 @@ package ethash
 
 import (
 	"encoding/binary"
+
 	"github.com/ethereum/go-ethereum/common/hexutil"
 )
