+++ consensus/ethash/ethash_test.go
@@ -16,146 +16,154 @@
 
 package ethash
 
-//
-//// Tests that ethash works correctly in test mode.
-//func TestTestMode(t *testing.T) {
-//	header := &types.BlockHeader{Height: 1, Difficulty: big.NewInt(100)}
-//
-//	ethash := NewTester(nil, false)
-//	defer ethash.Close()
-//
-//	results := make(chan *types.Block)
-//	err := ethash.Seal(nil, types.NewBlockWithHeader(header), results, nil)
-//	if err != nil {
-//		t.Fatalf("failed to seal block: %v", err)
-//	}
-//	select {
-//	case block := <-results:
-//		header.Witness = block.Header.Witness
-//		if err := ethash.VerifySeal(nil, header); err != nil {
-//			t.Fatalf("unexpected verification error: %v", err)
-//		}
-//	case <-time.NewTimer(time.Second).C:
-//		t.Error("sealing result timeout")
-//	}
-//}
-//
-//// This test checks that cache lru logic doesn't crash under load.
-//// It reproduces https://github.com/ethereum/go-ethereum/issues/14943
-//func TestCacheFileEvict(t *testing.T) {
-//	tmpdir, err := ioutil.TempDir("", "ethash-test")
-//	if err != nil {
-//		t.Fatal(err)
-//	}
-//	defer os.RemoveAll(tmpdir)
-//	e := New(Config{CachesInMem: 3, CachesOnDisk: 10, CacheDir: tmpdir, PowMode: ModeTest}, nil, false)
-//	defer e.Close()
-//
-//	workers := 8
-//	epochs := 100
-//	var wg sync.WaitGroup
-//	wg.Add(workers)
-//	for i := 0; i < workers; i++ {
-//		go verifyTest(&wg, e, i, epochs)
-//	}
-//	wg.Wait()
-//}
-//
-//func verifyTest(wg *sync.WaitGroup, e *Ethash, workerIndex, epochs int) {
-//	defer wg.Done()
-//
-//	const wiggle = 4 * epochLength
-//	r := rand.New(rand.NewSource(int64(workerIndex)))
-//	for epoch := 0; epoch < epochs; epoch++ {
-//		block := int64(epoch)*epochLength - wiggle/2 + r.Int63n(wiggle)
-//		if block < 0 {
-//			block = 0
-//		}
-//		header := &types.BlockHeader{Height: uint64(block), Difficulty: big.NewInt(100)}
-//		e.VerifySeal(nil, header)
-//	}
-//}
-//
-//func TestRemoteSealer(t *testing.T) {
-//	ethash := NewTester(nil, false)
-//	defer ethash.Close()
-//
-//	api := &API{ethash}
-//	if _, err := api.GetWork(); err != errNoMiningWork {
-//		t.Error("expect to return an error indicate there is no mining work")
-//	}
-//	header := &types.BlockHeader{Height: 1, Difficulty: big.NewInt(100)}
-//	block := types.NewBlockWithHeader(header)
-//	sealhash := ethash.SealHash(header)
-//
-//	// Push new work.
-//	results := make(chan *types.Block)
-//	ethash.Seal(nil, block, results, nil)
-//
-//	var (
-//		work [3]string
-//		err  error
-//	)
-//	if work, err = api.GetWork(); err != nil || work[0] != sealhash.Hex() {
-//		t.Error("expect to return a mining work has same hash")
-//	}
-//
-//	if res := api.SubmitWork(BlockNonce{}, sealhash, common.Hash{}); res {
-//		t.Error("expect to return false when submit a fake solution")
-//	}
-//	// Push new block with same block number to replace the original one.
-//	header = &types.BlockHeader{Height: 1, Difficulty: big.NewInt(1000)}
-//	block = types.NewBlockWithHeader(header)
-//	sealhash = ethash.SealHash(header)
-//	ethash.Seal(nil, block, results, nil)
-//
-//	if work, err = api.GetWork(); err != nil || work[0] != sealhash.Hex() {
-//		t.Error("expect to return the latest pushed work")
-//	}
-//}
-//
-//func TestHashRate(t *testing.T) {
-//	aHash, _ := common.HexToHash("a")
-//	bHash, _ := common.HexToHash("b")
-//	cHash, _ := common.HexToHash("c")
-//
-//	var (
-//		hashrate = []hexutil.Uint64{100, 200, 300}
-//		expect   uint64
-//
-//		ids      = []common.Hash{aHash, bHash, cHash}
-//	)
-//	ethash := NewTester(nil, false)
-//	defer ethash.Close()
-//
-//	if tot := ethash.Hashrate(); tot != 0 {
-//		t.Error("expect the result should be zero")
-//	}
-//
-//	api := &API{ethash}
-//	for i := 0; i < len(hashrate); i += 1 {
-//		if res := api.SubmitHashRate(hashrate[i], ids[i]); !res {
-//			t.Error("remote miner submit hashrate failed")
-//		}
-//		expect += uint64(hashrate[i])
-//	}
-//	if tot := ethash.Hashrate(); tot != float64(expect) {
-//		t.Error("expect total hashrate should be same")
-//	}
-//}
-//
-//func TestClosedRemoteSealer(t *testing.T) {
-//	ethash := NewTester(nil, false)
-//	time.Sleep(1 * time.Second) // ensure exit channel is listening
-//	ethash.Close()
-//
-//	api := &API{ethash}
-//	if _, err := api.GetWork(); err != errEthashStopped {
-//		t.Error("expect to return an error to indicate ethash is stopped")
-//	}
-//
-//	aHash, _ := common.HexToHash("a")
-//	if res := api.SubmitHashRate(hexutil.Uint64(100), aHash); res {
-//		t.Error("expect to return false when submit hashrate to a stopped ethash")
-//	}
-//}
+import (
+	"io/ioutil"
+	"math/big"
+	"math/rand"
+	"os"
+	"sync"
+	"testing"
+	"time"
+
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/seeleteam/go-seele/common"
+	"github.com/seeleteam/go-seele/core/types"
+	"github.com/stretchr/testify/assert"
+)
+
+// Tests that ethash works correctly in test mode.
+func TestTestMode(t *testing.T) {
+	header := &types.BlockHeader{Height: 1, Difficulty: big.NewInt(100)}
+
+	ethash := NewTester(nil, false)
+	defer ethash.Close()
+
+	results := make(chan *types.Block)
+	err := ethash.Seal(nil, types.NewBlockWithHeader(header), nil, results)
+	if err != nil {
+		t.Fatalf("failed to seal block: %v", err)
+	}
+	select {
+	case block := <-results:
+		header.Witness = block.Header.Witness
+		if err := ethash.VerifySeal(nil, header); err != nil {
+			t.Fatalf("unexpected verification error: %v", err)
+		}
+	case <-time.NewTimer(time.Second).C:
+		t.Error("sealing result timeout")
+	}
+}
+
+// This test checks that cache lru logic doesn't crash under load.
+// It reproduces https://github.com/ethereum/go-ethereum/issues/14943
+func TestCacheFileEvict(t *testing.T) {
+	tmpdir, err := ioutil.TempDir("", "ethash-test")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.RemoveAll(tmpdir)
+	e := New(Config{CachesInMem: 3, CachesOnDisk: 10, CacheDir: tmpdir, PowMode: ModeTest}, nil, false)
+	defer e.Close()
+
+	workers := 8
+	epochs := 100
+	var wg sync.WaitGroup
+	wg.Add(workers)
+	for i := 0; i < workers; i++ {
+		go verifyTest(&wg, e, i, epochs)
+	}
+	wg.Wait()
+}
+
+func verifyTest(wg *sync.WaitGroup, e *Ethash, workerIndex, epochs int) {
+	defer wg.Done()
+
+	const wiggle = 4 * epochLength
+	r := rand.New(rand.NewSource(int64(workerIndex)))
+	for epoch := 0; epoch < epochs; epoch++ {
+		block := int64(epoch)*epochLength - wiggle/2 + r.Int63n(wiggle)
+		if block < 0 {
+			block = 0
+		}
+		header := &types.BlockHeader{Height: uint64(block), Difficulty: big.NewInt(100)}
+		e.VerifySeal(nil, header)
+	}
+}
+
+func TestRemoteSealer(t *testing.T) {
+	ethash := NewTester(nil, false)
+	defer ethash.Close()
+
+	api := &API{ethash}
+	if _, err := api.GetWork(); err != errNoMiningWork {
+		t.Error("expect to return an error indicate there is no mining work")
+	}
+	header := &types.BlockHeader{Height: 1, Difficulty: big.NewInt(100)}
+	block := types.NewBlockWithHeader(header)
+	sealhash := ethash.SealHash(header)
+
+	// Push new work.
+	results := make(chan *types.Block)
+	ethash.Seal(nil, block, nil, results)
+
+	var (
+		work [3]string
+		err  error
+	)
+	if work, err = api.GetWork(); err != nil || work[0] != sealhash.Hex() {
+		t.Error("expect to return a mining work has same hash")
+	}
+
+	if res := api.SubmitWork(BlockNonce{}, sealhash, common.Hash{}); res {
+		t.Error("expect to return false when submit a fake solution")
+	}
+	// Push new block with same block number to replace the original one.
+	header = &types.BlockHeader{Height: 1, Difficulty: big.NewInt(1000)}
+	block = types.NewBlockWithHeader(header)
+	sealhash = ethash.SealHash(header)
+	ethash.Seal(nil, block, nil, results)
+
+	if work, err = api.GetWork(); err != nil || work[0] != sealhash.Hex() {
+		t.Error("expect to return the latest pushed work")
+	}
+}
+
+func TestHashRate(t *testing.T) {
+	var (
+		hashrate = []hexutil.Uint64{100, 200, 300}
+		expect   uint64
+
+		ids = []common.Hash{common.StringToHash("a"), common.StringToHash("b"), common.StringToHash("c")}
+	)
+	ethash := NewTester(nil, false)
+	defer ethash.Close()
+
+	if tot := ethash.Hashrate(); tot != 0 {
+		t.Error("expect the result should be zero")
+	}
+
+	api := &API{ethash}
+	for i := 0; i < len(hashrate); i += 1 {
+		if res := api.SubmitHashRate(hashrate[i], ids[i]); !res {
+			t.Error("remote miner submit hashrate failed")
+		}
+		expect += uint64(hashrate[i])
+	}
+
+	assert.Equal(t, ethash.Hashrate(), float64(expect), "expect total hashrate should be same")
+}
+
+func TestClosedRemoteSealer(t *testing.T) {
+	ethash := NewTester(nil, false)
+	time.Sleep(1 * time.Second) // ensure exit channel is listening
+	ethash.Close()
+
+	api := &API{ethash}
+	if _, err := api.GetWork(); err != errEthashStopped {
+		t.Error("expect to return an error to indicate ethash is stopped")
+	}
+
+	if res := api.SubmitHashRate(hexutil.Uint64(100), common.StringToHash("a")); res {
+		t.Error("expect to return false when submit hashrate to a stopped ethash")
+	}
+}
