+++ consensus/pow/engine.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/rcrowley/go-metrics"
 	"github.com/seeleteam/go-seele/consensus"
+	"github.com/seeleteam/go-seele/consensus/utils"
 	"github.com/seeleteam/go-seele/core/store"
 	"github.com/seeleteam/go-seele/core/types"
 	"github.com/seeleteam/go-seele/log"
@@ -63,15 +64,7 @@ func (engine *Engine) VerifyHeader(store store.BlockchainStore, header *types.Bl
 		return consensus.ErrBlockInvalidParentHash
 	}
 
-	if header.Height != parent.Height+1 {
-		return consensus.ErrBlockInvalidHeight
-	}
-
-	if header.CreateTimestamp.Cmp(parent.CreateTimestamp) < 0 {
-		return consensus.ErrBlockCreateTimeOld
-	}
-
-	if err = verifyDifficulty(parent, header); err != nil {
+	if err := utils.VerifyHeaderCommon(header, parent); err != nil {
 		return err
 	}
 
@@ -88,7 +81,7 @@ func (engine *Engine) Prepare(store store.BlockchainStore, header *types.BlockHe
 		return err
 	}
 
-	header.Difficulty = getDifficult(header.CreateTimestamp.Uint64(), parent)
+	header.Difficulty = utils.GetDifficult(header.CreateTimestamp.Uint64(), parent)
 
 	return nil
 }
@@ -129,15 +122,6 @@ func (engine *Engine) Seal(store store.BlockchainStore, block *types.Block, stop
 	return nil
 }
 
-func verifyDifficulty(parent *types.BlockHeader, header *types.BlockHeader) error {
-	difficult := getDifficult(header.CreateTimestamp.Uint64(), parent)
-	if header.Difficulty.Cmp(difficult) != 0 {
-		return consensus.ErrBlockDifficultInvalid
-	}
-
-	return nil
-}
-
 func verifyTarget(header *types.BlockHeader) error {
 	headerHash := header.Hash()
 	var hashInt big.Int
@@ -155,37 +139,4 @@ func verifyTarget(header *types.BlockHeader) error {
 // getMiningTarget returns the mining target for the specified difficulty.
 func getMiningTarget(difficulty *big.Int) *big.Int {
 	return new(big.Int).Div(maxUint256, difficulty)
-}
-
-// getDifficult adjust difficult by parent info
-func getDifficult(time uint64, parentHeader *types.BlockHeader) *big.Int {
-	// algorithm:
-	// diff = parentDiff + parentDiff / 2048 * max (1 - (blockTime - parentTime) / 10, -99)
-	// target block time is 10 seconds
-	parentDifficult := parentHeader.Difficulty
-	parentTime := parentHeader.CreateTimestamp.Uint64()
-	if parentHeader.Height == 0 {
-		return parentDifficult
-	}
-
-	big1 := big.NewInt(1)
-	big99 := big.NewInt(-99)
-	big2048 := big.NewInt(2048)
-
-	interval := (time - parentTime) / 10
-	var x *big.Int
-	x = big.NewInt(int64(interval))
-	x.Sub(big1, x)
-	if x.Cmp(big99) < 0 {
-		x = big99
-	}
-
-	var y = new(big.Int).Set(parentDifficult)
-	y.Div(parentDifficult, big2048)
-
-	var result = big.NewInt(0)
-	result.Mul(x, y)
-	result.Add(parentDifficult, result)
-
-	return result
 }
